chore: add cookie-policy page
