+++ dapp/src/app/cookie-policy/style.css
@@ -1,0 +1,24 @@
+.cookie-policy-page {
+    #skcm-cookie-library {
+        h4,
+        p,
+        table th,
+        table td {
+            @apply text-names-neutral-92 text-body-md;
+        }
+        h4 {
+            @apply mt-4;
+        }
+        p {
+            @apply mt-2;
+        }
+    }
+
+    #skcm-necessary-cookies-table {
+        @apply mt-2;
+    }
+
+    #skcm-cookie-library__preferences__button {
+        @apply rounded-full bg-iota-neutral-20 text-title-sm;
+    }
+}
