+++ dapp/src/app/globals.css
@@ -1,5 +1,6 @@
 @import '@iota/apps-ui-kit/styles';
 @import '../components/disclaimer/CookieDisclaimer.css';
+@import '../app/cookie-policy/style.css';
 
 @tailwind base;
 @tailwind components;
