+++ dapp/src/components/disclaimer/CookieDisclaimer.tsx
@@ -13,40 +13,36 @@ import { AMP_COOKIES_KEY } from '@/lib/utils/analytics/constants';
 
 import { PrivacyPolicyDialog } from '../dialogs/PrivacyPolicyDialog';
 
-export const configuration: SKCMConfiguration = {
-    disclaimer: {
-        title: undefined,
-        body: 'By using this website, you agree with our use of cookies. See details in the ',
-        policyText: 'Privacy Policy',
-        acceptButtonText: 'Close',
-        onPolicyClick: () => {
-            const current = new URL(window.location.href);
-            current.searchParams.set('modal', 'privacy_policy');
-            window.history.replaceState({}, '', current.toString());
-            window.dispatchEvent(new PopStateEvent('popstate'));
-        },
-    },
-    services: {
-        customNecessaryCookies: [
-            {
-                name: AMP_COOKIES_KEY,
-                purpose:
-                    'Flag indicating that Amplitude analytics cookies may be created after consent',
-                expiry: '1 year',
-                type: 'http',
-                showDisclaimerIfMissing: true,
-            },
-        ],
-    },
-    onAcceptCookies: () => {
-        onAmplitudeConsentAccepted();
-    },
-    onDeclineCookies: () => {
-        onAmplitudeConsentDeclined();
-    },
-};
-
 export function CookieDisclaimer() {
+    const configuration: SKCMConfiguration = {
+        disclaimer: {
+            title: undefined,
+            body: 'By using this website, you agree with our use of cookies. See details in the ',
+            policyText: 'Privacy Policy',
+            acceptButtonText: 'Close',
+            onPolicyClick: () => {
+                window.location.href = '/cookie-policy';
+            },
+        },
+        services: {
+            customNecessaryCookies: [
+                {
+                    name: AMP_COOKIES_KEY,
+                    purpose:
+                        'Flag indicating that Amplitude analytics cookies may be created after consent',
+                    expiry: '1 year',
+                    type: 'http',
+                    showDisclaimerIfMissing: true,
+                },
+            ],
+        },
+        onAcceptCookies: () => {
+            onAmplitudeConsentAccepted();
+        },
+        onDeclineCookies: () => {
+            onAmplitudeConsentDeclined();
+        },
+    };
     return (
         <>
             <CookieManager configuration={configuration} />
